I FINALLY FIXED COLLISIONS

+++ src/player/CharAnim.h
@@ -8,7 +8,7 @@ class CharAnim : public Animation {
     public:
         CharAnim(bool repeat=true);
         virtual void Update(float dt);
-        void Draw(float x, float y, int spriteWidth, int spriteHeight, float xScale=1, float yScale=1, SDL_RendererFlip = SDL_FLIP_NONE);
+        void Draw(int x, int y, int spriteWidth, int spriteHeight, float xScale=1, float yScale=1, SDL_RendererFlip = SDL_FLIP_NONE);
         void SetProps(std::string textureID, int spriteRow, int frameCount, int speed);
     protected:
         int m_Speed;
@@ -17,3 +17,4 @@ class CharAnim : public Animation {
         std::string m_textureID;
 };
 #endif
+
